Moved all modals in separate dialogs folder
