--- src/app/modules/OpenshiftStreams/components/CreateInstanceModal/index.ts
@@ -1,1 +1,0 @@
-export * from './CreateInstanceModal';
