--- src/app/modules/OpenshiftStreams/components/DeleteInstanceModal/index.ts
@@ -1,1 +1,0 @@
-export * from './DeleteInstanceModal';
