+++ src/app/modules/OpenshiftStreams/components/index.ts
@@ -1,4 +1,2 @@
-export * from './CreateInstanceModal';
 export * from './StreamsTableView';
-export * from './DeleteInstanceModal';
 export * from './InstanceDrawer';
