+++ src/app/modules/OpenshiftStreams/dialogs/CreateInstance/CreateInstance.css
