+++ src/app/modules/OpenshiftStreams/dialogs/CreateInstance/CreateInstance.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen, waitFor, act } from '@testing-library/react';
 import { I18nextProvider } from 'react-i18next';
 import userEvent from '@testing-library/user-event';
-import { CreateInstanceModal, CreateInstanceModalProps } from './CreateInstanceModal';
+import { CreateInstance } from './CreateInstance';
 import i18nForTest from '../../../../../../test-utils/i18n';
 import { AuthContext } from '@app/auth/AuthContext';
 
@@ -32,7 +32,7 @@ jest.mock('../../../openapi/api', () => {
   };
 });
 
-const setupRender = (props: CreateInstanceModalProps) => {
+const setupRender = (props: any) => {
   render(
     <I18nextProvider i18n={i18nForTest}>
       <AuthContext.Provider
@@ -41,14 +41,14 @@ const setupRender = (props: CreateInstanceModalProps) => {
           getUsername: () => Promise.resolve('api_kafka_service'),
         }}
       >
-        <CreateInstanceModal {...props} />
+        <CreateInstance {...props} />
       </AuthContext.Provider>
     </I18nextProvider>
   );
 };
 
-describe('<CreateInstanceModal/>', () => {
-  const props: CreateInstanceModalProps = {
+describe('<CreateInstance/>', () => {
+  const props: any = {
     createStreamsInstance: true,
     mainToggle: true,
     cloudProviders: [
