+++ src/app/modules/OpenshiftStreams/dialogs/CreateInstance/DrawerPanelContentInfo.tsx
