+++ src/app/modules/OpenshiftStreams/dialogs/CreateInstance/index.ts
@@ -1,0 +1,1 @@
+export * from './CreateInstance';
