+++ src/app/modules/OpenshiftStreams/dialogs/DeleteInstance/index.ts
@@ -1,0 +1,1 @@
+export * from './DeleteInstance';
