+++ src/app/modules/OpenshiftStreams/dialogs/index.ts
@@ -1,0 +1,2 @@
+export * from './CreateInstance';
+export * from './DeleteInstance';
