--- src/app/modules/ServiceAccounts/components/ResetServiceAccountModal/index.ts
@@ -1,1 +1,0 @@
-export * from './ResetServiceAccountModal';
