+++ src/app/modules/ServiceAccounts/components/index.ts
@@ -1,4 +1,1 @@
-export * from './CreateServiceAccountModal';
-export * from './DeleteServiceAccountModal';
-export * from './ResetServiceAccountModal';
 export * from './ServiceAccountsTableView';
