+++ src/app/modules/ServiceAccounts/dialogs/CreateServiceAccount/index.ts
@@ -1,0 +1,1 @@
+export * from './CreateServiceAccount';
