+++ src/app/modules/ServiceAccounts/dialogs/DeleteServiceAccount/index.ts
@@ -1,0 +1,1 @@
+export * from './DeleteServiceAccount';
