+++ src/app/modules/ServiceAccounts/dialogs/ResetServiceAccount/index.ts
@@ -1,0 +1,1 @@
+export * from './ResetServiceAccount';
