+++ src/app/modules/ServiceAccounts/dialogs/index.ts
@@ -1,0 +1,3 @@
+export * from './CreateServiceAccount';
+export * from './DeleteServiceAccount';
+export * from './ResetServiceAccount';
